Project Organization
Moved Some Files Around
Include new Styles Class
Moved some functions in Menu.cpp

Console Class Adjustments

Initialization Changes
Colors Struct -> Enumerator

Include Missing Changes

forgot to reinclude dllmain
Adjusted g_Console->printdbg to default to WHITE if no color is given
