+++ EROverlay.vcxproj
@@ -185,19 +185,20 @@
     <ClInclude Include="libs\Minhook\table32.h" />
     <ClInclude Include="libs\Minhook\table64.h" />
     <ClInclude Include="libs\Minhook\trampoline.h" />
-    <ClInclude Include="MENU\ChrIns.hpp" />
-    <ClInclude Include="MENU\ChrModules.hpp" />
-    <ClInclude Include="MENU\Common.hpp" />
-    <ClInclude Include="MENU\Console.hpp" />
-    <ClInclude Include="MENU\D3DRenderer.hpp" />
-    <ClInclude Include="MENU\Game.hpp" />
-    <ClInclude Include="MENU\GameDataMan.hpp" />
-    <ClInclude Include="MENU\GameEnums.hpp" />
-    <ClInclude Include="MENU\Hooking.hpp" />
-    <ClInclude Include="MENU\Memory.hpp" />
-    <ClInclude Include="MENU\Menu.hpp" />
-    <ClInclude Include="MENU\Vector.hpp" />
-    <ClInclude Include="MENU\WorldCharMan.hpp" />
+    <ClInclude Include="MENU\include\ChrIns.hpp" />
+    <ClInclude Include="MENU\include\ChrModules.hpp" />
+    <ClInclude Include="MENU\include\Common.hpp" />
+    <ClInclude Include="MENU\include\Console.hpp" />
+    <ClInclude Include="MENU\include\D3DRenderer.hpp" />
+    <ClInclude Include="MENU\include\framework.h" />
+    <ClInclude Include="MENU\include\Game.hpp" />
+    <ClInclude Include="MENU\include\GameDataMan.hpp" />
+    <ClInclude Include="MENU\include\GameEnums.hpp" />
+    <ClInclude Include="MENU\include\Hooking.hpp" />
+    <ClInclude Include="MENU\include\Memory.hpp" />
+    <ClInclude Include="MENU\include\Menu.hpp" />
+    <ClInclude Include="MENU\include\Tools\Vector.hpp" />
+    <ClInclude Include="MENU\include\WorldCharMan.hpp" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="libs\ImGui\imgui.cpp" />
@@ -212,17 +213,17 @@
     <ClCompile Include="libs\Minhook\hde64.c" />
     <ClCompile Include="libs\Minhook\hook.c" />
     <ClCompile Include="libs\Minhook\trampoline.c" />
-    <ClCompile Include="MENU\ChrIns.cpp" />
-    <ClCompile Include="MENU\ChrModules.cpp" />
-    <ClCompile Include="MENU\Common.cpp" />
-    <ClCompile Include="MENU\Console.cpp" />
-    <ClCompile Include="MENU\D3DRenderer.cpp" />
     <ClCompile Include="MENU\dllmain.cpp" />
-    <ClCompile Include="MENU\Game.cpp" />
-    <ClCompile Include="MENU\GameDataMan.cpp" />
-    <ClCompile Include="MENU\Hooking.cpp" />
-    <ClCompile Include="MENU\Menu.cpp" />
-    <ClCompile Include="MENU\WorldCharMan.cpp" />
+    <ClCompile Include="MENU\src\ChrIns.cpp" />
+    <ClCompile Include="MENU\src\ChrModules.cpp" />
+    <ClCompile Include="MENU\src\Common.cpp" />
+    <ClCompile Include="MENU\src\Console.cpp" />
+    <ClCompile Include="MENU\src\D3DRenderer.cpp" />
+    <ClCompile Include="MENU\src\Game.cpp" />
+    <ClCompile Include="MENU\src\GameDataMan.cpp" />
+    <ClCompile Include="MENU\src\Hooking.cpp" />
+    <ClCompile Include="MENU\src\Menu.cpp" />
+    <ClCompile Include="MENU\src\WorldCharMan.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
