--- MENU/ChrIns.cpp
@@ -1,1 +1,0 @@
-#include "ChrIns.hpp"
