+++ MENU/include/WorldCharMan.hpp
