+++ MENU/src/ChrIns.cpp
@@ -1,0 +1,1 @@
+#include "../include/ChrIns.hpp"
