+++ MENU/src/Common.cpp
@@ -1,4 +1,4 @@
-#include "Common.hpp"
+#include "../include/Common.hpp"
 /// <summary>
 /// Clear Bytes
 /// </summary>
